finregistry-pipeline: Reorganize project structure
Move all files except the following to risteys_pipeline.finngen:
- .gitignore
- Dockerfile
- main.wdl
- README
- requirements.txt

Fix imports to match the new project structure.
Setup the pipeline as a python module.
Setup logging.
Replace README.adoc with README.md and update the README.
