+++ .gitignore
@@ -1,9 +1,3 @@
-# Python
-__pycache__/
-venv/
-.ipynb_checkpoints/
-/notebooks/
-
 # Javascript
 node_modules/
