+++ pipeline/setup.py
@@ -1,0 +1,3 @@
+from setuptools import setup
+
+setup(name = 'risteys_pipeline', version = '0.1')
